requirements.txt
fastapi
uvicorn
langchain-core
langchain-community
langchain-openai
python-dotenv
pydantic
requests
langgraph